Refactor: reorg, EntryList -> Locations
