+++ src/components/mobile/MobileLayout.js
@@ -11,11 +11,11 @@ import { EditLocationForm } from '../form/EditLocation'
 import { EditReviewForm } from '../form/EditReview'
 import { LocationForm } from '../form/LocationForm'
 import { ReviewForm } from '../form/ReviewForm'
+import ListPage from '../list/ListPage'
 import MapPage from '../map/MapPage'
 import SettingsPage from '../settings/SettingsPage'
 import { zIndex } from '../ui/GlobalStyle'
 import { PageTabs, TabList, TabPanels } from '../ui/PageTabs'
-import ListPage from './ListPage'
 import Tabs from './Tabs'
 import TopBarSwitch from './TopBarSwitch'
